use local submodules except for necessary branches
The submodules are first and foremost for getting at the documentation
of the various plugins, and for being able to edit that documentation so
as to be able to push changes, but we can also use the code in them for
bit-docs as plugins by use of file: paths, so to avoid remote npm.

For those plugin repos that have necessary branches that have not yet
been merged into master, we pull from GitHub instead of checking out
that branch in the local submodule so as to leave the submodule
available for the explicit purpose of editing documentation.

If the necessary branches were still in development, we could clone them
locally anywhere and point the file: to that location as needed, thus
having one copy of the plugin repo for editing documentation and another
for developing the code of the plugin.